==> 01_generate_spectrograms.ipynb <==
# Step 1: Generating spectrograms from audio files (and some cleaning)

This script creates spectrograms from a list of input audio files.
#### The following minimal structure and files are required in the project directory:

    ├── info_file.csv      <- A ";"-separated csv file containing metadata about the calls
    ├── audio
    │   ├── call_1.wav     <- call files (Mono, not stereo sound!)
    │   ├── call_2.wav     <- 
    │   ├── call_3.wav     <- 
    │   └── ...            <- 
    ├── data               <- empty directory for output files    

#### The following structure is required for info_file.csv:
(must contain at least the two columns: "filename" and "label"). If labels are completely unknown, there should still be a label column with some NA identifiers (e..g "unknown")

    | filename   | label   | ...    |  .... 
    -----------------------------------------
    | call_1.wav | alarm   |  ...   |  ....   
    | call_2.wav | contact |  ...   |  ....  
    | ...        |  ...    |  ...   |  ....   

#### The following files are generated in this script:

    ├── data               
    │   ├── df.pkl <- pickled pandas dataframe with metadata, raw_audio and spectrograms

## Import statements, constants and functions

import pandas as pd
import numpy as np
import pickle
import os
from pathlib import Path
import umap

from audio_functions import generate_mel_spectrogram, read_wavfile
from preprocessing_functions import calc_zscore, pad_spectro

BANDPASS_FILTER = False  # should bandpass-filtered spectrograms be generated?
MEDIAN_SUB = False  # should median-subtracted spectrograms be generated (reduce impulse noise)?
STRETCH = False    # should time-stretched spectrograms be generated (all stretched to max. duration in dataset)?

LABEL_COL = "label"     # name of column that contains labels
NA_DESCRIPTORS = [0, np.nan, "NA", "na", "not available", # values that indicate "no label"
                  "None", "Unknown", "unknown", None, ""]     # add your NA descriptor if not yet in list
                                                     
NEW_NA_INDICATOR = "unknown" # all vocalizations without label will be relabelled as "unknown"

P_DIR = os.getcwd() # --> project directory default: your current working directory

AUDIO_IN = os.path.join(os.path.sep, P_DIR, 'audio') # --> audio directory, contains audio (.wav) files
DATA = os.path.join(os.path.sep, P_DIR, 'data') # --> empty data directory, output files will be put here


# Check if directories are present

if not os.path.isdir(AUDIO_IN):
    print("No audio directory found")

if not os.path.isdir(DATA):
    os.mkdir(DATA)

## 1. Read in files

df = pd.read_csv('info_file.csv', sep=";", index_col=[0])

df

#### Check if all audio files are in AUDIO_IN directory

audiofiles = df['filename'].values
files_in_audio_directory = os.listdir(AUDIO_IN)

missing_files = list(set(audiofiles) - set(files_in_audio_directory))
if len(missing_files)>0:
    print(len(missing_files), "files with no matching audio in audio folder: ", missing_files)

audio_filepaths = [os.path.join(os.path.sep, AUDIO_IN,x) for x in audiofiles]

## 2. Adding audio (and samplerate) to dataframe

raw_audio,samplerate_hz = map(list,zip(*[read_wavfile(x) for x in audio_filepaths]))

df['raw_audio'] = raw_audio
df['samplerate_hz'] = samplerate_hz

# Removing NA rows

nrows = df.shape[0]
df.dropna(subset=['raw_audio'], inplace=True)
print("Dropped ", nrows-df.shape[0], " rows due to missing/failed audio")

## 3. Removing very long calls

It's advisable to remove very long calls, as all calls will be zero-padded to the maximum duration in the dataset.

# Extract duration of calls
df['duration_s'] = [x.shape[0] for x in df['raw_audio']]/df['samplerate_hz']

[Can be helpful to plot the distribution to find a good cutoff]

#%matplotlib inline
#n, bins, patches = plt.hist(df['duration_s'])

In our case, dataset was already cleaned and all calls were between 0-0.5s, so no need to remove long calls. Set MIN_DUR and MAX_DUR to values that make sense for your dataset.

MIN_DUR = 0  # --> minimum duration of calls in seconds
MAX_DUR = 0.5 # --> maximum duration of calls in seconds

print("Dropped ", df.loc[df['duration_s']<MIN_DUR,:].shape[0], "rows below ", MIN_DUR, "s (min_dur)")
df = df.loc[df['duration_s']>=MIN_DUR,:]
print("Dropped ", df.loc[df['duration_s']>MAX_DUR,:].shape[0], "rows above ", MAX_DUR, "s (max_dur)")
df = df.loc[df['duration_s']<=MAX_DUR,:]

## 4. Generate mel-spectrograms

In this step, spectrograms are generated from audio files via short-time fourier transformation. Spectrograms capture the frequency components of a signal over time. A spectrogram is a 2D matrix, where each value represents the signal intensity in a specific time (columns) and frequency bin (row). In this case, the frequency axis of the spectrograms are also Mel-transformed (a logarithmic scale) and signal intensity is expressed on a Decibel scale.

The following parameters define how the spectrograms are computed. You can leave this as default or choose your own parameters.

### 4.1. Set spectrogramming parameters

N_MELS = 40 # --> number of mel bins (usually 20-40)
            # The frequency bins are transformed to this number of logarithmically spaced mel bins.

FFT_WIN = 0.03 # --> length of audio chunk when applying STFT in seconds
               # FFT_WIN * samplerate = number of audio datapoints that go in one fft (=n_fft)

FFT_HOP = FFT_WIN/8 # --> hop_length in seconds
                    # FFT_HOP * samplerate = n of audio datapoints between successive ffts (=hop_length)

WINDOW = 'hann' # --> name of window function
                # each frame of audio is windowed by a window function. We use the window function 'hanning',  

FMIN = 0 # --> lower bound for frequency (in Hz) when generating Mel filterbank
FMAX = int(np.min(df['samplerate_hz'])/2) #--> upper bound for frequency (in Hz) when generating Mel filterbank
                                                 # this is set to 0.5 times the samplerate (-> Nyquist rule)
                                                 # If input files have different samplerates, the lowest samplerate is used
                                                 # to ensure all spectrograms have the same frequency resolution.

Save these parameters as they will be needed for all subsequent scripts

lines = ['N_MELS = '+str(N_MELS),
         'FFT_WIN = '+str(FFT_WIN),
         'FFT_HOP = '+str(FFT_HOP),
         'WINDOW = "'+str(WINDOW)+'"',
         'FMIN = '+str(FMIN),
         'FMAX = '+str(FMAX)]

with open(os.path.join(os.path.sep, P_DIR, 'spec_params.py'), 'w') as f:
    for line in lines:
        f.write(line)
        f.write('\n')

### 4.2. Generate spectrograms

spectrograms = df.apply(lambda row: generate_mel_spectrogram(row['raw_audio'],
                                                                    row['samplerate_hz'],
                                                                    N_MELS,
                                                                    WINDOW,
                                                                    FFT_WIN,
                                                                    FFT_HOP,
                                                                    FMAX), 
                               axis=1)


df['spectrograms'] = spectrograms

# Removing NA rows

nrows = df.shape[0]
df.dropna(subset=['spectrograms'], inplace=True)
print("Dropped ", nrows-df.shape[0], " rows due to failed spectrogram generation")

## [Optional: 5. Generate denoised files (median subtraction):]

if MEDIAN_SUB:
    df['denoised_spectrograms'] = [(spectrogram - np.median(spectrogram, axis=0)) for spectrogram in df['spectrograms']]

## [Optional : 6. Generate denoised files (bandpass filter):]

if BANDPASS_FILTER:
    from scipy.signal import butter, lfilter

    # Bandpass filters for calculating audio intensity
    LOWCUT = 300.0
    HIGHCUT = 3000.0

    # Butter bandpass filter implementation:
    # from https://scipy-cookbook.readthedocs.io/items/ButterworthBandpass.html

    def butter_bandpass(lowcut, highcut, fs, order=5):
        nyq = 0.5 * fs
        low = lowcut / nyq
        high = highcut / nyq
        b, a = butter(order, [low, high], btype='band')
        return b, a

    def butter_bandpass_filter(data, lowcut, highcut, fs, order=5):
        b, a = butter_bandpass(lowcut, highcut, fs, order=order)
        y = lfilter(b, a, data)
        return y
    
    df['filtered_audio'] = df.apply(lambda row: butter_bandpass_filter(row['raw_audio'], 
                                                                   LOWCUT, 
                                                                   HIGHCUT, 
                                                                   row['samplerate_hz'],
                                                                   order=6),
                               axis=1)

    filtered_spectrograms = df.apply(lambda row: generate_mel_spectrogram(row['filtered_audio'],
                                                                        row['samplerate_hz'],
                                                                        N_MELS,
                                                                        WINDOW,
                                                                        FFT_WIN,
                                                                        FFT_HOP,
                                                                        FMAX), 
                                   axis=1)


    df['filtered_spectrograms'] = filtered_spectrograms

## [Optional: 7. Generate stretched spectrograms (all stretched to equal length)]

if STRETCH:
    MAX_DURATION = np.max(df['duration_s'])

    df['stretched_spectrograms'] = df.apply(lambda row: generate_stretched_mel_spectrogram(row['raw_audio'],
                                                                                           row['samplerate_hz'],
                                                                                           row['duration_s'],
                                                                                           N_MELS,
                                                                                           WINDOW,
                                                                                           FFT_WIN,
                                                                                           FFT_HOP,
                                                                                           MAX_DURATION),
                                            axis=1)

## 8. Clean labels

Transform all labels into strings and label all calls without label as "unknown"

df['original_label'] = df[LABEL_COL] #  original labels are saved in "original_label" column

df['label'] = ["unknown" if x in NA_DESCRIPTORS else x for x in df[LABEL_COL]]  # "unknown" for all NA labels
labels = df['label'].fillna(NEW_NA_INDICATOR) # double-check
df['label'] = labels.astype(str) # transform to strings and save in df as "label" column

## 9. Save dataframe

df.to_pickle(os.path.join(os.path.sep, DATA, 'df.pkl'))
